realiza estrutura organizacional do projeto. ----- falta testar operacionalidade. Adiciona também o docker compose para comunicar com o banco de dados através do mesmo.
