+++ docker-compose.yml
@@ -1,0 +1,15 @@
+version: "3.1"
+
+services:
+    # Serviço que irá rodar o banco de dados
+  db:
+    container_name: movieProjectApi_db
+    image: mysql:8.0.32
+    restart: always
+    # Lista de portas mapeadas de dentro para fora do container
+    # na notação porta_de_fora:porta_de_dentro
+    ports:
+      # Expõe a porta padrão do banco de dados
+      - 3306:3306
+    environment:
+      MYSQL_ROOT_PASSWORD: password
