+++ pom.xml
@@ -14,7 +14,7 @@
 	<name>MovieProjectApi</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
-		<java.version>20</java.version>
+		<java.version>17</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
