+++ src/main/resources/application.properties
@@ -1,6 +1,7 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/movie_project_db
+spring.datasource.url=jdbc:mysql://db:3306/movie_project_db?createDatabaseIfNotExist=true
 spring.datasource.username=root
 spring.datasource.password=password
+spring.main.allow-circular-references=true
 
 spring.jpa.show-sql=true
 spring.jpa.generate-ddl=true
